[S] Moved Collision detection out of `Player` to clean up code
